Restructure build pipeline. Split up in project and templates
