+++ .teamcity/Templates/LintTemplate.kt
@@ -1,0 +1,33 @@
+package Templates
+
+import jetbrains.buildServer.configs.kotlin.DslContext
+import jetbrains.buildServer.configs.kotlin.Template
+import jetbrains.buildServer.configs.kotlin.buildSteps.script
+
+object LintTemplate : Template({
+    name = "LintTemplate"
+
+    detectHangingBuilds = false
+
+    vcs {
+        root(DslContext.settingsRoot, "+:. => imod-python")
+
+        cleanCheckout = true
+    }
+
+    steps {
+        script {
+            name = "Static code analysis"
+            id = "Static_code_analysis"
+            workingDir = "imod-python"
+            scriptContent = """
+                    pixi run --environment default --frozen lint 
+                """.trimIndent()
+            formatStderrAsError = true
+        }
+    }
+
+    requirements {
+        equals("env.OS", "Windows_NT")
+    }
+})
